# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ChatRequest" : {
        "properties" : {
          "history" : {
            "items" : {
              "$ref" : "#/components/schemas/ChatTurn"
            },
            "type" : "array"
          },
          "observed" : {
            "minLength" : 1,
            "type" : "string"
          },
          "observer" : {
            "minLength" : 1,
            "type" : "string"
          },
          "question" : {
            "minLength" : 1,
            "type" : "string"
          },
          "reasoningLevel" : {
            "type" : "string"
          },
          "responseFormat" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "session" : {
            "type" : "string"
          }
        },
        "required" : [ "observed", "observer", "question" ],
        "type" : "object"
      },
      "ChatResponse" : {
        "properties" : {
          "answer" : {
            "type" : "string"
          },
          "iterations" : {
            "format" : "int32",
            "type" : "integer"
          },
          "stoppedAtLimit" : {
            "type" : "boolean"
          },
          "toolCalls" : {
            "items" : {
              "$ref" : "#/components/schemas/ToolCallResponse"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "ChatTurn" : {
        "properties" : {
          "content" : {
            "minLength" : 1,
            "type" : "string"
          },
          "role" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "content", "role" ],
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "EventResponse" : {
        "properties" : {
          "actor" : {
            "type" : "string"
          },
          "afterContent" : {
            "type" : "string"
          },
          "beforeContent" : {
            "type" : "string"
          },
          "conclusionId" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "detail" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "event" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ToolCallResponse" : {
        "properties" : {
          "arguments" : {
            "type" : "string"
          },
          "failed" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "description" : "A token from `POST /v1/tokens`, in four nesting scopes: admin, workspace, peer, session. Each operation below names the narrowest one that may call it.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Pair-scoped memory. Every fact belongs to a directed (observer, observed) pair: what bob remembers about alice is not what alice remembers about herself, and no route crosses that line.\n\nThree tiers of read: `context` assembles a session's summary and recent messages within a token budget, `recall` ranks conclusions by six fused signals, and `chat` runs an agentic loop over both when the answer needs more than one lookup.",
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "aimon-memory",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/v1/workspaces/{workspace}/chat" : {
      "post" : {
        "description" : "**Requires a `peer` token or wider.**",
        "operationId" : "chat",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChatRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChatResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No model provider is configured."
          }
        },
        "summary" : "Tier 2: answer a question with tools over memory",
        "tags" : [ "chat" ]
      }
    },
    "/v1/workspaces/{workspace}/chat/stream" : {
      "post" : {
        "description" : "**Requires a `peer` token or wider.**",
        "operationId" : "stream",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChatRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/event-stream" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "A stream of `delta` events carrying answer text, then a single `done` event."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No model provider is configured."
          }
        },
        "summary" : "Tier 2, streamed over SSE",
        "tags" : [ "chat" ]
      }
    },
...
    "/v1/workspaces/{workspace}/conclusions/{id}/events" : {
      "get" : {
        "description" : "**Requires a `peer` token or wider.**",
        "operationId" : "history",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EventResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Read a conclusion's audit trail",
        "tags" : [ "conclusions" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "description" : "Relative to wherever this service is deployed.",
    "url" : "/"
  } ],
...
}